version Beta 1.3 Need to active Notification service

+++ src/main/java/com/ahmadfahd/Services/RatingServices.java
@@ -8,6 +8,6 @@ public interface RatingServices {
     List<RatingDTO> getAllRatings();
     RatingDTO findById(Long rateid);
     void addRating(int rate, Long uId, Long eId);
-    float findRateAvg(Long uid);
+    double findRateAvg(Long uid);
 
 }
